+++ apm-sniffer/apm-sdk-plugin/httpclient-5.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/httpclient/v5/HttpClientDoExecuteInterceptor.java
@@ -95,7 +95,7 @@ public abstract class HttpClientDoExecuteInterceptor implements InstanceMethodsA
         if (host == null) {
             return true;
         }
-        return isExcludedPort(host.getPort());
+        return isExcludedPort(port(host));
     }
 
     /**